Changes Done As Mention Food menu should be list as per outlet ( here outlet id text and comma separator like 1,2,3) only display item those belongs to the login user outlet id Order grid: Price and total should be with .00 discount popup value should be reset while clicking on cancel service charge popup value should be reset while clicking on cancel replace textbox with the label for the discount and service charge The total amount in the grid and all other amounts below grid should be in the same line with right align Remove order No text Can we remove the "btnAll" and adding from the list dynamic in code or joins in a query? pl make sure for the Isactive data can show  true for all master data Receipt number should be with six-digit like 000005,000015,000568 Add customer name like 00015 [ Jagdish Mackwan ] [order Number ]+ [customer name] Apply virtual keyboard on all textbox as per login page Menu list: click open popup the open food menu in the grid with search Menu name, Price, category, (Double click on grid add item to sale grid) SalesInvoiceNumber entry when bill generate

+++ RocketPOS.Model/SaleItemModel.cs
@@ -9,9 +9,9 @@ namespace RocketPOS.Model
         public int CustomerOrderItemId { get; set; }
         public string FoodMenuId { get; set; }
         public string Product { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public decimal Qty { get; set; }
         public decimal Discount { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
     }
 }
